Creating cart module instead of database, data, models modules
