+++ feature/pizza/src/main/java/ru/evgenykuzakov/pizza/data/mapper/CatalogMapper.kt
@@ -1,6 +1,6 @@
 package ru.evgenykuzakov.pizza.data.mapper
 
-import ru.evgenykuzakov.models.pizza.type.SizeType
+import ru.evgenykuzakov.cart.domain.model.SizeType
 import ru.evgenykuzakov.pizza.data.model.CatalogDto
 import ru.evgenykuzakov.pizza.domain.model.Catalog
 
@@ -9,5 +9,5 @@ fun CatalogDto.toDomain() = Catalog(
     description = description,
     img = img,
     name = name,
-    priceFrom = sizes.map { it.toDomain() }.first{ it.type == SizeType.Small }.price
+    priceFrom = sizes.map { it.toDomain() }.first{ it.type == ru.evgenykuzakov.cart.domain.model.SizeType.Small }.price
 )
