+++ feature/pizza/src/main/java/ru/evgenykuzakov/pizza/data/mapper/SizeMapper.kt
@@ -1,9 +1,9 @@
 package ru.evgenykuzakov.pizza.data.mapper
 
-import ru.evgenykuzakov.models.pizza.type.SizeType
+import ru.evgenykuzakov.cart.domain.model.SizeType
 import ru.evgenykuzakov.pizza.data.model.SizeDto
 
-fun SizeDto.toDomain() = ru.evgenykuzakov.models.pizza.Size(
+fun SizeDto.toDomain() = ru.evgenykuzakov.cart.domain.model.Size(
     price = price,
-    type = SizeType.fromString(type)
+    type = ru.evgenykuzakov.cart.domain.model.SizeType.fromString(type)
 )
