+++ feature/pizza/src/main/java/ru/evgenykuzakov/pizza/data/mapper/ToppingMapper.kt
@@ -1,10 +1,10 @@
 package ru.evgenykuzakov.pizza.data.mapper
 
-import ru.evgenykuzakov.models.pizza.type.IngredientType
+import ru.evgenykuzakov.cart.domain.model.IngredientType
 import ru.evgenykuzakov.pizza.data.model.ToppingDto
 
-fun ToppingDto.toDomain() = ru.evgenykuzakov.models.pizza.Topping(
+fun ToppingDto.toDomain() = ru.evgenykuzakov.cart.domain.model.Topping(
     img = img,
     price = price,
-    type = IngredientType.fromString(type)
+    type = ru.evgenykuzakov.cart.domain.model.IngredientType.fromString(type)
 )
