+++ settings.gradle.kts
@@ -30,5 +30,5 @@ include(":components")
 include(":components:theme")
 include(":components:ui")
 include(":feature:pizza")
-include(":shared:database")
 include(":shared:models")
+include(":shared:cart")
