+++ shared/cart/src/main/AndroidManifest.xml
