+++ shared/cart/src/main/java/ru/evgenykuzakov/cart/data/database/model/DoughEntity.kt
@@ -1,0 +1,9 @@
+package ru.evgenykuzakov.cart.data.database.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class DoughEntity(
+    val price: Int,
+    val type: String
+)
