+++ shared/cart/src/main/java/ru/evgenykuzakov/cart/data/database/model/IngredientEntity.kt
@@ -1,0 +1,10 @@
+package ru.evgenykuzakov.cart.data.database.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class IngredientEntity(
+    val img: String,
+    val price: Int,
+    val type: String
+)
