+++ shared/cart/src/main/java/ru/evgenykuzakov/cart/data/database/model/PizzaEntity.kt
@@ -1,0 +1,14 @@
+package ru.evgenykuzakov.cart.data.database.model
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "cart")
+data class PizzaEntity(
+    @PrimaryKey(autoGenerate = true) val id: String,
+    val name: String,
+    val img: String,
+    val toppings: List<ToppingEntity> = emptyList(),
+    val size: SizeEntity,
+    val dough: DoughEntity
+)
