+++ shared/cart/src/main/java/ru/evgenykuzakov/cart/data/database/model/SizeEntity.kt
@@ -1,0 +1,9 @@
+package ru.evgenykuzakov.cart.data.database.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SizeEntity(
+    val price: Int,
+    val type: String
+)
