+++ shared/cart/src/main/java/ru/evgenykuzakov/cart/data/database/model/ToppingEntity.kt
@@ -1,0 +1,10 @@
+package ru.evgenykuzakov.cart.data.database.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ToppingEntity(
+    val img: String,
+    val price: Int,
+    val type: String
+)
