+++ shared/cart/src/main/java/ru/evgenykuzakov/cart/data/mapper/CatalogMapper.kt
@@ -1,0 +1,13 @@
+package ru.evgenykuzakov.cart.data.mapper
+
+import ru.evgenykuzakov.cart.data.database.model.PizzaEntity
+import ru.evgenykuzakov.cart.domain.model.Pizza
+
+fun PizzaEntity.toDomain() = Pizza(
+    id = id,
+    img = img,
+    name = name,
+    toppings = toppings.map { it.toDomain() },
+    dough = dough.toDomain(),
+    size = size.toDomain()
+)
