+++ shared/cart/src/main/java/ru/evgenykuzakov/cart/di/DataModule.kt
@@ -1,0 +1,19 @@
+package ru.evgenykuzakov.cart.di
+
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import ru.evgenykuzakov.cart.data.repository.CartRepositoryImpl
+import ru.evgenykuzakov.cart.domain.repository.CartRepository
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class DataModule {
+
+    @Binds
+    internal abstract fun bindsLocalUsersRepository(
+        impl: CartRepositoryImpl,
+    ): CartRepository
+
+}
