+++ shared/cart/src/main/java/ru/evgenykuzakov/cart/domain/model/Ingredient.kt
@@ -1,0 +1,9 @@
+package ru.evgenykuzakov.cart.domain.model
+
+import ru.evgenykuzakov.cart.domain.model.type.IngredientType
+
+data class Ingredient(
+    val img: String,
+    val price: Int,
+    val type: IngredientType
+)
