+++ shared/cart/src/main/java/ru/evgenykuzakov/cart/domain/model/Size.kt
@@ -1,0 +1,8 @@
+package ru.evgenykuzakov.cart.domain.model
+
+import ru.evgenykuzakov.cart.domain.model.type.SizeType
+
+data class Size(
+    val price: Int,
+    val type: SizeType
+)
