+++ shared/cart/src/main/java/ru/evgenykuzakov/cart/domain/model/type/DoughType.kt
@@ -1,0 +1,15 @@
+package ru.evgenykuzakov.cart.domain.model.type
+
+sealed class DoughType(val name: String) {
+    data object Thin : DoughType("THIN")
+    data object Thick : DoughType("THICK")
+
+    companion object {
+        fun fromString(value: String): DoughType =
+            when (value.uppercase()) {
+                "THIN" -> Thin
+                "THICK" -> Thick
+                else -> throw IllegalArgumentException("Unknown dough type: $value")
+            }
+    }
+}
