+++ shared/cart/src/main/java/ru/evgenykuzakov/cart/domain/repository/CartRepository.kt
@@ -1,0 +1,9 @@
+package ru.evgenykuzakov.cart.domain.repository
+
+import ru.evgenykuzakov.cart.domain.model.Pizza
+
+interface CartRepository {
+
+    suspend fun getCart(): List<Pizza>
+
+}
