--- shared/database/src/main/java/ru/evgenykuzakov/database/model/PizzaEntity.kt
@@ -1,8 +1,0 @@
-package ru.evgenykuzakov.database.model
-
-data class PizzaEntity(
-    val id: String,
-    val name: String,
-    val img: String,
-
-)
