--- shared/models/src/main/java/ru/evgenykuzakov/models/pizza/Dough.kt
@@ -1,8 +1,0 @@
-package ru.evgenykuzakov.models.pizza
-
-import ru.evgenykuzakov.models.pizza.type.DoughType
-
-data class Dough(
-    val price: Int,
-    val type: DoughType
-)
