--- shared/models/src/main/java/ru/evgenykuzakov/models/pizza/Ingredient.kt
@@ -1,9 +1,0 @@
-package ru.evgenykuzakov.models.pizza
-
-import ru.evgenykuzakov.models.pizza.type.IngredientType
-
-data class Ingredient(
-    val img: String,
-    val price: Int,
-    val type: IngredientType
-)
