--- shared/models/src/main/java/ru/evgenykuzakov/models/pizza/Size.kt
@@ -1,8 +1,0 @@
-package ru.evgenykuzakov.models.pizza
-
-import ru.evgenykuzakov.models.pizza.type.SizeType
-
-data class Size(
-    val price: Int,
-    val type: SizeType
-)
