--- shared/models/src/main/java/ru/evgenykuzakov/models/pizza/Topping.kt
@@ -1,9 +1,0 @@
-package ru.evgenykuzakov.models.pizza
-
-import ru.evgenykuzakov.models.pizza.type.IngredientType
-
-data class Topping(
-    val img: String,
-    val price: Int,
-    val type: IngredientType
-)
