--- shared/models/src/main/java/ru/evgenykuzakov/models/pizza/type/SizeType.kt
@@ -1,17 +1,0 @@
-package ru.evgenykuzakov.models.pizza.type
-
-sealed class SizeType(val name: String) {
-    data object Small : SizeType("SMALL")
-    data object Medium : SizeType("MEDIUM")
-    data object Large : SizeType("LARGE")
-
-    companion object {
-        fun fromString(value: String): SizeType =
-            when (value.uppercase()) {
-                "SMALL" -> Small
-                "MEDIUM" -> Medium
-                "LARGE" -> Large
-                else -> throw IllegalArgumentException("Unknown size type: $value")
-            }
-    }
-}
